refactored auth-component and put inside each component
